refactor: Nested Class virou DTO com Nasted Class
